add event names
